+++ Cargo.toml
@@ -14,4 +14,5 @@ syn = { version = "2", default-features = false, features = ["full", "parsing"] 
 prettyplease = "0.2"
 proc-macro2 = "1"
 sha256 = "1.4"
+serde_json = "1.0.107"
